#215 refactor real-time tickers using new model
