+++ build.gradle
@@ -276,6 +276,26 @@ project(":exchanges-connector") {
     }
 }
 
+project(":tickers") {
+
+    apply plugin: "org.springframework.boot"
+
+    springBoot {
+        mainClassName = 'fund.cyber.markets.TickersApplication'
+    }
+
+    dependencies {
+        compile project(":common")
+        compile project(":common-kafka")
+        compile project(":cassandra-service")
+
+        compile 'org.springframework.boot:spring-boot-starter'
+        compile 'org.springframework.kafka:spring-kafka'
+
+        testCompile 'org.springframework.boot:spring-boot-starter-test'
+    }
+}
+
 project(":tickers-batch") {
 
     apply plugin: "org.springframework.boot"
@@ -292,7 +312,7 @@ project(":tickers-batch") {
         compile 'org.springframework.boot:spring-boot-starter'
         compile 'org.ehcache:ehcache'
 
-        testCompile("org.springframework.boot:spring-boot-starter-test")
+        testCompile 'org.springframework.boot:spring-boot-starter-test'
     }
 }
